(chore) Reduces circular references. Moves field config map into get field config to avoid hot reload issues related to circulars. Moves form utils into utils directory.
