+++ containers/mosaic/src/components/Field/FormFieldTextEditor/NodeForm/NodeFormLink.tsx
@@ -4,7 +4,8 @@ import React, { useMemo, useCallback } from "react";
 import type { FieldDef } from "@root/components/Field";
 import type { NodeFormTypeProps, TextEditorUpdateLink } from "../FormFieldTextEditorTypes";
 
-import Form, { useForm } from "@root/components/Form";
+import Form from "@root/components/Form/Form";
+import { useForm } from "@root/components/Form/useForm/useForm";
 import { LinkOpen } from "./LinkOpen";
 import { NodeFormFooter } from "./NodeFormFooter";
