+++ containers/mosaic/src/components/Form/index.ts
@@ -2,4 +2,3 @@ export { default } from "./Form";
 export * from "./Form";
 export * from "./FormTypes";
 export * from "./useForm";
-export * from "./formUtils";
