+++ containers/mosaic/src/components/MosaicContext/MosaicContext.ts
@@ -1,6 +1,6 @@
 import { createContext } from "react";
-import { defaulti18n } from "../../i18n";
-import { i18n } from "i18next";
+import { defaulti18n } from "../../i18n/internal";
+import type { i18n } from "i18next";
 
 export interface MosaicContextProps {
 	i18n: i18n;
