+++ containers/mosaic/src/utils/form/cleanValue.ts
@@ -1,0 +1,9 @@
+export const cleanValue = (value: any) => {
+	if (value === "" || (Array.isArray(value) && value.length === 0)) {
+		return undefined;
+	}
+
+	return value;
+};
+
+export default cleanValue;
