+++ containers/mosaic/src/utils/form/defaultResolver.ts
@@ -1,0 +1,7 @@
+import cleanValue from "./cleanValue";
+
+export function defaultResolver(value: any) {
+	return { internalValue: value, value: cleanValue(value) };
+}
+
+export default defaultResolver;
